Move Topic and Queue into a namespace. Rename toX methods to use service name
