+++ packages/punchcard/lib/stream/index.ts
@@ -2,7 +2,7 @@ export * from './collector';
 export * from './delivery-stream';
 export * from './glue';
 export * from './kinesis';
-export * from './queue';
 export * from './sink';
+export * from './sns';
+export * from './sqs';
 export * from './stream';
-export * from './topic';
